Phuong: Chỉnh sửa code
Chỉnh sửa load theo config id, chạy 1 process
